+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - `properties = graphGetEdgeProperties(graphId, fromNodeId, toNodeId)`
 - `isSuccess = graphRemoveEdge(graphId, fromNodeId, toNodeId)`
 - `path = graphFindPath(graphId, startNodeId, endNodeId, [skipNodeIds], [skipNodeTypes], [skipEdgeTypes])`
+- `isSuccess = graphPrecomputePathsToNode(graphId, nodeId, [skipNodeTypes], [skipEdgeTypes])`
 - `nodeId = graphFindClosestNode(graphId, origin)`
 - `edgeInfo = graphFindClosestEdge(graphId, origin)`
 - Fixed graph node type initialization (affects `skipNodeTypes` filtering in path finding).